requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorzero==2.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
HX711 @ git+https://github.com/gandalf15/HX711.git@a2b72a0b44fc200112ecb25e11b60a5ee4ba7d56#subdirectory=HX711_Python3
idna==3.6
pycrypto==2.6.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
RPi.GPIO==0.7.1
smbus==1.1.post2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
